# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "CabaPro Árbitros API (Node Gateway)",
        "version": "1.0.0",
        "description": "API Node/Express que consume la API SpringBoot y expone funcionalidades del rol Árbitro"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "summary": "Registro de árbitro (proxy a Spring)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "Login de árbitro (proxy a Spring)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/arbitros/dashboard": {
            "get": {
                "summary": "Dashboard del árbitro (proxy por defecto)",
                "parameters": [
                    {
                        "name": "mode",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "proxy",
                                "redirect"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RegisterRequest": {
                "type": "object",
                "properties": {
                    "correo": {
                        "type": "string"
                    },
                    "contrasena": {
                        "type": "string"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "especialidad": {
                        "type": "string"
                    },
                    "escalafon": {
                        "type": "string"
                    }
                },
                "required": [
                    "correo",
                    "contrasena",
                    "nombre"
                ]
            },
            "LoginRequest": {
                "type": "object",
                "properties": {
                    "correo": {
                        "type": "string"
                    },
                    "contrasena": {
                        "type": "string"
                    }
                },
                "required": [
                    "correo",
                    "contrasena"
                ]
            },
...
        }
    }
}